V2 - Beta Phase
The most important part, the mp3 player has been done. The soundtrack operation is still a bit robust, but otherwise it's functional

+++ Javafy/src/concept/Music.java
@@ -3,16 +3,16 @@ package src.concept;
 public class Music {
     private final String musicName;
     private final String filePath;
-    private final int duration;
+    private final double duration;
     private final String artist;
-    Music(String nombre,String path,int duration,String artist) {
+    Music(String nombre,String path,double duration,String artist) {
         this.musicName = nombre;
         this.filePath = path;
         this.duration = duration;
         this.artist = artist;
     }
 
-    public int getDuration() {
+    public double getDuration() {
         return duration;
     }
